Increased GCC bitchiness settings
Bitchiness settings used: -Wall -Wno-unused-function -Wno-pointer-sign
and -Werror

-Werror because warnings will be missed during compilation.

Increased warnings to pick up on a lot of stuff.

-Wno-unused-functions because of some stuff that's partially ported
-Wno-pointer-sign because there are a lot of pointer sign things, but
they are almost all unsigned char * <-> char * for dealing with buffers.

-pedantic is not used becaues something in the CMSIS headers makes it mad,
so it will probably never be able to be added.

Ideally both of the -Wno- flags will be able to removed once porting is
done, but that won't happen until the Great Refactoring, where we will
switch everything over to stdint.h types and come up with a common naming
scheme for everything.